Make it possible to plot aelo curves via the oq plot command
