+++ openquake/calculators/postproc/compute_rtgm.py
@@ -37,12 +37,8 @@ Useful abbreviations:
 """
 import io
 import logging
-import json
 import numpy as np
 import pandas as pd
-import matplotlib as mpl
-from matplotlib import pyplot as plt
-from scipy import interpolate
 from scipy.interpolate import RegularGridInterpolator
 try:
     import rtgmpy
@@ -56,7 +52,9 @@ from openquake.baselib import hdf5
 from openquake.hazardlib.imt import from_string
 from openquake.hazardlib.calc.mean_rates import to_rates
 from openquake.calculators import postproc
-from openquake.calculators.extract import get_info
+from openquake.commands.plot import (
+    plot_mean_hcurves_afe_rtgm, plot_disagg_by_src, plot_governing_mce,
+    _find_fact_maxC)
 
 DLL_df = pd.read_csv(io.StringIO('''\
 imt,A,B,BC,C,CD,D,DE,E
@@ -103,40 +101,6 @@ def norm_imt(imt):
     Normalize the imt string to the USGS format, for instance SA(1.1) -> SA1P1
     """
     return imt.replace('(', '').replace(')', '').replace('.', 'P')
-
-
-f1 = interpolate.interp1d([0.2, 1], [1.1, 1.3])
-f2 = interpolate.interp1d([1, 5], [1.3, 1.5])
-f3 = interpolate.interp1d([0.2, 1], [1.2, 1.25])
-f4 = interpolate.interp1d([0.2, 1], [1.1, 1.3])
-
-
-def _find_fact_maxC(T, code):
-    # Find the factor to convert to maximum component based on
-    # ASCE7-16 and ASCE7-22
-    if code == 'ASCE7-16':
-        if T == 0:
-            fact_maxC = 1.
-        elif T <= 0.2:
-            fact_maxC = 1.1
-        elif T <= 1:
-            fact_maxC = f1(T)
-        elif T <= 5:
-            fact_maxC = f2(T)
-        else:
-            fact_maxC = 1.5
-    elif code == 'ASCE7-22':
-        if T == 0:
-            fact_maxC = 1.
-        elif T <= 0.2:
-            fact_maxC = 1.2
-        elif T <= 1:
-            fact_maxC = f3(T)
-        elif T <= 10:
-            fact_maxC = f4(T)
-        else:
-            fact_maxC = 1.5
-    return fact_maxC
 
 
 def calc_rtgm_df(hcurves, sitecol, oq):
@@ -378,293 +342,6 @@ def get_asce41(dstore, mce, facts):
     return asce41
 
 
-def _get_label(imt):
-    imtlab = imt if imt == 'PGA' else imt.replace(')', 's)')
-    comp = 'Geom. mean' if imt == 'PGA' else 'Max. comp.'
-    return imtlab + ' - ' + comp
-
-
-def plot_meanHCs_afe_RTGM(imls, AFE, RTGM, afe_RTGM, imts):
-    """
-    :param imls: intensity measure levels
-    :param AFE: hazard curves in terms of annual frequency of exceedance
-    :param RTGM: risk targeted ground motion
-    :afe_RTGM: annual frequency of exceedance associated to RTGM
-    :imts: intensity measure types of interest
-    :returns: figure of hazard curves
-    """
-    plt.figure(figsize=(12, 9))
-    plt.rcParams.update({'font.size': 16})
-    colors = mpl.colormaps['viridis'].reversed()._resample(3)
-    patterns = ['-', '--', ':']
-    for i, imt in enumerate(imts):
-        lab = _get_label(imt)
-        plt.loglog(imls[i], AFE[i], color=colors(i), linestyle=patterns[i],
-                   label=lab, linewidth=3, zorder=1)
-        # plot the label only once but it must be at the end of the legend
-        if imt == imts[-1]:
-            plt.loglog([RTGM[i]], [afe_RTGM[i]], 'ko',
-                       label='Probabilistic MCE',  linewidth=2,
-                       markersize=10, zorder=3)
-        else:
-            plt.loglog([RTGM[i]], [afe_RTGM[i]], 'ko',
-                       linewidth=2, markersize=10, zorder=3)
-        plt.loglog([np.min(imls[i]), RTGM[i]], [afe_RTGM[i], afe_RTGM[i]],
-                   'darkgray', linestyle='--', linewidth=1)
-        plt.loglog([RTGM[i], RTGM[i]], [0, afe_RTGM[i]], 'darkgray',
-                   linestyle='--', linewidth=1)
-
-    plt.grid('both')
-    plt.legend(fontsize=13)
-    plt.xlabel('Acceleration (g)', fontsize=20)
-    plt.ylabel('Annual frequency of exceedance', fontsize=20)
-    plt.legend(loc="best", fontsize='16')
-    plt.ylim([10E-6, 1.1])
-    plt.xlim([0.01, 4])
-    bio = io.BytesIO()
-    plt.savefig(bio, format='png', bbox_inches='tight')
-    plt.clf()
-    return Image.open(bio)
-
-
-def _find_afe_target(imls, afe, sa_target):
-    # find the target afe (or poe) for a given acceleration
-    if len(imls) != len(afe):
-        afe.extend([1E-15] * (len(imls) - len(afe)))
-    f = interpolate.interp1d(np.log(imls), np.log(afe))
-    afe_target = np.exp(f(np.log(sa_target)))
-    return afe_target
-
-
-# TODO: this is horrible code to be removed
-def disaggr_by_src(dstore):
-    # get info : specific to disagg by src
-    df = dstore['mean_rates_by_src'].to_dframe().set_index('src_id')
-    grouped_m = df.groupby(['src_id', 'site_id', 'imt']).agg(
-        {"value": list}).reset_index()
-    # remove the sources that aren't contributing at all to the hazard
-    mask = grouped_m.value.apply(lambda x: sum(x) > 0)
-    gm = grouped_m[mask].reset_index()
-    grouped_2 = gm.groupby(['imt', 'src_id']).agg(
-        {"value": np.array}).reset_index()
-    total_poe = []
-    for wp in grouped_2.value.values:
-        wsp = []
-        if isinstance(wp, list):
-            total_poe.append(wp)
-        else:
-            for wp_i in wp:
-                wsp.append(wp_i)
-            total_poe.append([sum(t) for t in np.array(wsp).T])
-    grouped_2['poes'] = total_poe
-    return grouped_2
-
-
-def _find_sources(df, imtls, imts, rtgm_probmce, mean_hcurve, dstore):
-
-    fig, ax = plt.subplots(3, figsize=(8, 15))
-
-    # identify the sources that have a contribution > than fact (here 10%) of
-    # the largest contributor;
-    fact = 0.1
-
-    for m, imt in enumerate(imts):
-        out_contr_all = []
-        fig1, ax1 = plt.subplots()
-
-        dms = df[(df['imt'] == imt)]
-        # annual frequency of exceedance:
-        T = from_string(imt).period
-        f = 0 if imt == 0.0 else _find_fact_maxC(T, 'ASCE7-16')
-        imls_o = imtls[imt]
-        imls = [iml*f for iml in imls_o]
-        # have to compute everything for max comp. and for geom. mean
-        RTGM = rtgm_probmce[m]
-        RTGM_o = rtgm_probmce[m]/f
-        afe_target = _find_afe_target(imls, mean_hcurve[m], RTGM)
-        afe_target_o = _find_afe_target(imls_o, mean_hcurve[m], RTGM_o)
-
-        # populate 3-panel plot
-        ax[m].loglog(imls, mean_hcurve[m], 'k', label=_get_label(imt),
-                     linewidth=2, zorder=3)
-        ax[m].loglog([np.min(imls), RTGM], [afe_target, afe_target], 'k--',
-                     linewidth=2, zorder=3)
-        ax[m].loglog([RTGM, RTGM], [0, afe_target], 'k--', linewidth=2,
-                     zorder=3)
-        ax[m].loglog([RTGM], [afe_target], 'ko', label='Probabilistic MCE',
-                     linewidth=2, zorder=3)
-        # populate individual plots
-        ax1.loglog(imls_o, mean_hcurve[m], 'k', label=imt + ' - Geom. mean',
-                   linewidth=2, zorder=3)
-        ax1.loglog([np.min(imls_o), RTGM_o], [afe_target_o, afe_target_o],
-                   'k--', linewidth=2, zorder=3)
-        ax1.loglog([RTGM_o, RTGM_o], [0, afe_target_o], 'k--', linewidth=2,
-                   zorder=3)
-        ax1.loglog([RTGM_o], [afe_target_o], 'ko', label='Probabilistic MCE',
-                   linewidth=2, zorder=3)
-
-        # poes from dms are now rates
-        for ind, (afes, src) in enumerate(zip(dms.poes, dms.src_id)):
-            # get contribution at target level for that source
-            afe_uhgm = _find_afe_target(imls, afes, rtgm_probmce[m])
-            # get % contribution of that source
-            contr_source = afe_uhgm/afe_target
-            out_contr_all.append(contr_source * 100)
-
-        # identify contribution of largest contributor, make color scale
-        largest_contr = np.max(out_contr_all)
-        sample = sum(out_contr_all > fact*largest_contr)
-        viridis = mpl.colormaps['viridis'].reversed()._resample(sample)
-
-        # find and plot the sources, highlighting the ones that contribute more
-        # than 10% of largest contributor
-        # use j to only add the "other sources" label once
-        # use i to cycle through the colors for the major source contributors
-        i = j = 0
-        for ind, (afes, src) in enumerate(zip(dms.poes, dms.src_id)):
-            # pad to have the same length of imls and afes
-            afe_pad = afes + [0] * (len(imls) - len(afes))
-            # if it's not a big contributor, plot in silver
-            if out_contr_all[ind] <= fact*largest_contr:
-                if j == 0:
-                    ax[m].loglog(imls, afe_pad, 'silver', linewidth=0.7,
-                                 label='other sources')
-                    ax1.loglog(imls_o, afe_pad, 'silver', linewidth=0.7,
-                               label='other source')
-                    j += 1
-                else:
-                    ax[m].loglog(imls, afe_pad, 'silver', linewidth=0.7)
-                    ax1.loglog(imls_o, afe_pad, 'silver', linewidth=0.7)
-            # if it is, plot in color
-            else:
-                ax[m].loglog(imls, afe_pad, c=viridis(i), label=str(src))
-                ax1.loglog(imls_o, afe_pad, c=viridis(i), label=str(src))
-                i += 1
-        # populate subplot - maximum component
-        ax[m].grid('both')
-        ax[m].set_xlabel(imt+' (g)', fontsize=16)
-        ax[m].set_ylabel('Annual Freq. Exceedance', fontsize=16)
-        ax[m].legend(loc='center left', bbox_to_anchor=(1, 0.5), fontsize='13')
-        ax[m].set_ylim([10E-6, 1.1])
-        ax[m].set_xlim([0.01, 4])
-
-        # populate single imt plots - geometric mean
-        ax1.grid('both')
-        ax1.set_xlabel(imt+' (g)', fontsize=16)
-        ax1.set_ylabel('Annual Freq. Exceedance', fontsize=16)
-        ax1.legend(loc='center left', bbox_to_anchor=(1, 0.5), fontsize='13')
-        ax1.set_ylim([10E-6, 1.1])
-        ax1.set_xlim([0.01, 4])
-
-        # save single imt plot
-        bio1 = io.BytesIO()
-        fig1.savefig(bio1, format='png', bbox_inches='tight')
-        # keep these in webui until we finish checks and have a command line
-        # exporter, then we can change the name to _{imt} and they will not
-        # appear in the webui
-        dstore[f'png/disagg_by_src-{imt}.png'] = Image.open(bio1)
-
-    # save triple plot
-    bio = io.BytesIO()
-    fig.savefig(bio, format='png', bbox_inches='tight')
-    logging.info('Storing png/disagg_by_src.png')
-    dstore['png/disagg_by_src-All-IMTs.png'] = Image.open(bio)
-
-
-def plot_governing_mce(dstore, imtls, low_haz):
-    """
-    :param dstore: the datastore
-    :param imtls: intensity measure types and levels
-    :param low_haz: boolean specifying if the hazard is lower than DLLs
-    :returns: image of governing MCE
-    """
-    js = dstore['asce7'][()].decode('utf8')
-    dic = json.loads(js)
-    MCEr = [dic['PGA'], dic['SS'], dic['S1']]
-    T = [from_string(imt).period for imt in imtls]
-
-    limit_det = [0.5, 1.5, 0.6]
-    # presenting as maximum component -> do not need conversion facts
-    RTGM = dstore.read_df('rtgm')['ProbMCE']
-    plt.figure(figsize=(8, 6))
-    plt.rcParams.update({'font.size': 15})
-    plt.plot(T, limit_det, 'kx', markersize=15, label='DLL', linewidth=1)
-    plt.plot(T[0], RTGM[0], 'bX', markersize=12, label='$PGA_{GM}$',
-             linewidth=3)
-    plt.plot(T[1:], RTGM[1:], 'bs', markersize=12,
-             label='$S_{S,RT}$ and $S_{1,RT}$', linewidth=3)
-    if low_haz:
-        plt.ylim([0, np.max([RTGM, MCEr, limit_det]) + 0.2])
-    else:
-        MCEr_det = [dic['PGA_84th'], dic['SS_84th'], dic['S1_84th']]
-        plt.plot(T[0], MCEr_det[0], 'c^', markersize=10, label='$PGA_{84th}$',
-                 linewidth=3)
-        plt.plot(T[1:], MCEr_det[1:], 'cd', markersize=10,
-                 label='$S_{S,84th}$ and $S_{1,84th}$', linewidth=3)
-        plt.ylim([0, np.max([RTGM,  MCEr, MCEr_det, limit_det]) + 0.2])
-    plt.scatter(T[0], MCEr[0], s=200, label='Governing $MCE_G$',
-                linewidth=2, facecolors='none', edgecolors='r')
-    plt.scatter(T[1:], MCEr[1:], s=200, marker='s',
-                label='Governing $MCE_R$', linewidth=2,
-                facecolors='none', edgecolors='r')
-    plt.grid('both')
-    plt.ylabel('Spectral Acceleration (g)', fontsize=20)
-    plt.xlabel('Period (s)', fontsize=20)
-    plt.legend(loc="upper right", fontsize='13')
-    plt.xlim([-0.02, 1.2])
-    bio = io.BytesIO()
-    plt.savefig(bio, format='png', bbox_inches='tight')
-    plt.clf()
-    return Image.open(bio)
-
-
-def plot_curves(dstore, low_haz):
-    """
-    Calls functions to plot curves
-
-    :param dstore: the datastore
-    :param low_haz: boolean specifying if the hazard is lower than DLLs
-    """
-    dinfo = get_info(dstore)
-    # site is always 0 for a single-site calculation
-    # get imls and imts, make arrays
-    imtls = dinfo['imtls']
-    # separate imts and imls
-    AFE, afe_target, imls = [], [], []
-    for imt in IMTS:
-        # get periods and factors for converting btw geom mean and
-        # maximum component
-        T = from_string(imt).period
-        f = 0 if imt == 0.0 else _find_fact_maxC(T, 'ASCE7-16')
-        imls.append([im*f for im in imtls[imt]])
-    # get rtgm ouptut from the datastore
-    rtgm_df = dstore.read_df('rtgm')
-    # get the IML for the 2475 RP
-    rtgm_probmce = rtgm_df['ProbMCE']
-    # get investigation time
-    window = dinfo['investigation_time']
-    # get hazard curves, put into rates
-    mean_hcurve = dstore['hcurves-stats'][0, 0]  # shape(M, L1)
-    for m, hcurve in enumerate(mean_hcurve):
-        AFE.append(to_rates(hcurve, window, minrate=1E-12))
-        # get the AFE of the iml that will be disaggregated for each IMT
-        afe_target.append(_find_afe_target(
-            imls[m], AFE[m], rtgm_probmce[m]))
-    # make plot
-    img = plot_meanHCs_afe_RTGM(imls, AFE, rtgm_probmce, afe_target, IMTS)
-    logging.info('Storing png/hcurves.png')
-    dstore['png/hcurves.png'] = img
-
-    if low_haz:
-        img = plot_governing_mce(dstore, imtls, low_haz=True)
-    else:
-        df = disaggr_by_src(dstore)
-        _find_sources(df, imtls, IMTS, rtgm_probmce, mean_hcurve, dstore)
-        img = plot_governing_mce(dstore, imtls, low_haz=False)
-    logging.info('Storing png/governing_mce.png')
-    dstore['png/governing_mce.png'] = img
-
-
 def main(dstore, csm):
     """
     :param dstore: datastore with the classical calculation
@@ -700,7 +377,9 @@ def main(dstore, csm):
         if Image is None:  # missing PIL
             logging.warning('Missing module PIL: skipping plotting curves')
         else:
-            plot_curves(dstore, low_haz=True)
+            # make_aelo_plots(dstore, low_haz=True)
+            plot_mean_hcurves_afe_rtgm(dstore, update_dstore=True)
+            plot_governing_mce(dstore, update_dstore=True)
         return
 
     mag_dist_eps, sigma_by_src = postproc.disagg_by_rel_sources.main(
@@ -722,4 +401,7 @@ def main(dstore, csm):
     if Image is None:  # missing PIL
         logging.warning('Missing module PIL: skipping plotting curves')
     else:
-        plot_curves(dstore, low_haz=False)
+        # make_aelo_plots(dstore, low_haz=False)
+        plot_mean_hcurves_afe_rtgm(dstore, update_dstore=True)
+        plot_disagg_by_src(dstore, update_dstore=True)
+        plot_governing_mce(dstore, update_dstore=True)
